Lexisort implemented and quasi-tested
Sorting by substrings implemented and tested using substring print
function.  Not intense testing, but working perfectly.  Substring
generation method updated to not generate past end of schedule.
Algorithm updated to add more to the substring if 0 or 1 in substring
(so compare method which starts at t+1 has something to compare based
on)

Next: Debug flag to turn debugging on and off